Accomodate models with >32767 unique senone sequences without crashing, although there are still some residual issues
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@8341 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/mdef.h
@@ -98,7 +98,7 @@ typedef struct {
  * \brief Triphone information, including base phones as a subset.  For the latter, lc, rc and wpos are non-existent.
  */
 typedef struct {
-    int16 ssid;			/**< State sequence (or senone sequence) ID, considering the
+    int32 ssid;			/**< State sequence (or senone sequence) ID, considering the
 				   n_emit_state senone-ids are a unit.  The senone sequences
 				   themselves are in a separate table */
     int32 tmat;			/**< Transition matrix id */
